+++ examples/example.py
@@ -1,5 +1,6 @@
 import forecastingapi as fapi
 import pandas as pd 
+from time import time
 
 if __name__ == "__main__":
     
@@ -10,30 +11,45 @@ if __name__ == "__main__":
 
     # 2 - get a token 
     token = fapi.get_token(username = "user20231122@example.com",
-                                   password = "pwd")
-    print(token)
+                                   password = "pwd") # choose a better password
 
-    ## 3 - get forecast with prediction interval
+    ## 3 - get forecast with prediction interval (for a file or a url)
     path_to_file = 'https://raw.githubusercontent.com/Techtonique/datasets/main/time_series/univariate/AirPassengers.csv'
-    #path_to_file = './datasets/AirPassengers.csv'
+    path_to_file2 = './datasets/AirPassengers.csv' 
      
+    print("\n Example 1 --------------- \n")
+    start = time() 
     res_get_forecast = fapi.get_forecast(file=path_to_file, 
-    token=token)
+    token=token, method = "theta")
+    print(f"Elapsed: {time() - start} seconds \n")
 
-    print(pd.DataFrame(res_get_forecast))
+    print(f"mean forecast: \n {pd.DataFrame(res_get_forecast['averages'], columns=['date', 'mean'])}\n")
+    print(f"prediction intervals: \n {pd.DataFrame(res_get_forecast['ranges'], columns=['date', 'lower', 'upper'])}")
 
-    res_get_forecast2 = fapi.get_forecast(file=path_to_file, 
-    token=token, start_training = 2, n_training = 7, h = 4, level = 90)
+    print("\n Example 2 --------------- \n")
+    start = time()
+    res_get_forecast2 = fapi.get_forecast(file=path_to_file2, 
+    token=token, start_training = 2, n_training = 7, h = 10, level = 90)
+    print(f"Elapsed: {time() - start} seconds")
 
-    print(pd.DataFrame(res_get_forecast2))
+    print(f"mean forecast: \n {pd.DataFrame(res_get_forecast2['averages'], columns=['date', 'mean'])}\n")
+    print(f"prediction intervals: \n {pd.DataFrame(res_get_forecast2['ranges'], columns=['date', 'lower', 'upper'])}")
 
+    print("\n Example 3 --------------- \n")
+    start = time()
     res_get_forecast3 = fapi.get_forecast(file=path_to_file, 
     token=token, date_formatting="ms",
-    start_training = 2, n_training = 7, h = 4, level = 90)
+    start_training = 2, n_training = 7, h = 10, level = 90)
+    print(f"Elapsed: {time() - start} seconds")
 
-    print(pd.DataFrame(res_get_forecast3))
+    print(f"mean forecast: \n {pd.DataFrame(res_get_forecast3['averages'], columns=['date', 'mean'])}\n")
+    print(f"prediction intervals: \n {pd.DataFrame(res_get_forecast3['ranges'], columns=['date', 'lower', 'upper'])}")
 
-    res_get_forecast4 = fapi.get_forecast(file=path_to_file, 
-    token=token, method = "prophet")
+    print("\n Example 4 --------------- \n")
+    start = time()
+    res_get_forecast4 = fapi.get_forecast(file=path_to_file2, 
+    token=token, method = "prophet", h=10)
+    print(f"Elapsed: {time() - start} seconds \n")
 
-    print(pd.DataFrame(res_get_forecast4))
+    print(f"mean forecast: \n {pd.DataFrame(res_get_forecast4['averages'], columns=['date', 'mean'])}\n")
+    print(f"prediction intervals: \n {pd.DataFrame(res_get_forecast4['ranges'], columns=['date', 'lower', 'upper'])}")