+++ CHANGES.md
@@ -1,3 +1,10 @@
+# version 0.2.0 (2023-11-22)
+
+- use a csv available online (through url, not just on Desktop) in requests
+- import a few csv data in the package 
+- better error message when token has expired or file format is not supported 
+- run examples: `make run-examples`
+
 # version 0.1.0 (2022-11-22)
 
 - initial version with Theta forecasting, mean forecasting, Random walk forecasting, and FB Prophet forecasting